Cast Streaming: Add RTP Payload Type hacks for AndroidTV.
While working on the audio part of the Receiver demo, it was discovered
that the Chrome Sender will always put a fixed RTP payload type of 127
in the RTP packet headers. This was meant to be a workaround for some
buggy AndroidTV Receiver implementations. This change adds
RtpPayloadType enums to document this, and patches the code to permit
the InspectPacketForRouting() function to recognize packets with this
payload type as a valid RTP packet.

Bug: openscreen:56
Change-Id: I1fcc6e5e509ef18e76726cba9edebb77d1e82d76