Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > Shared vs unique records:
> > https://tools.ietf.org/html/rfc6762#section-2
> > 
> > Also:
> > https://tools.ietf.org/html/rfc6762#section-8.4
> > https://tools.ietf.org/html/rfc6762#section-10.2
> 
> Thanks, that was helpful!
> 
> Can you shed some light on how a responder knows if it is responding with a shared or unique record?

Responder has to go through a probing step to establish uniqueness. This means responder must send a query for the record it's registering to determine if it's a unique one.

> Can you give an example of a shared record during normal mDNS use?
PTR records in DNS-SD are the reason why this is necessary.
All PTR records answer the same question with the service name used a domain name.
All answers point to a different target domain, which is a name of the instance of the service.

Patch-set: 9
Reviewer: Gerrit User 1336186 <1336186@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0