Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (4 comments)
> > 
> > Ok.  Having the ctor set the singleton is a little bit of a footgun, as normally a ctor only side effects the object you're constructing and its collaborators, not global state.  Is the concern here that the caller would construct a PlatformClientImpl and forget to call SetInstance()?
> 
> Yuri's original comment is as follows:
> "Instead of client-side code having to manually call Create() and ShutDown(), how about just making the instance a local, scoped variable:
> 
>   PlatformClientPosix platform_client(microseconds(50), microseconds(50));
> ...then, no matter how the method returns, the dtor will be called automatically.
> 
> This also means getting rid of the PlatformClientPosix::Create() and Shutdown() methods, and then making the ctor/dtor public. Simpler! :)"
> 
> I am in favor of leaving the PlatformClient as it currently is, but wanted this detail hashed out before proceeding as Yuri has strong opinions about it and wants it resolved before checking in the code to update our demos

I'm okay with having the dtor clear the singleton as that ties the lifetime of the object and the state of the singleton pointer.

The fact that the object is a singleton is an important part of the API and I'm reluctant to hide that as an implementation detail in the constructor though.  Otherwise you get unexpected behavior when constructing the second instance.  It would be better to not allow the second instance to be constructed at all, by making the constructor private, and accessing construction through a static method.  That eliminates the possibility of misuse.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>