Update patch set 1
Patch Set 1:

> The fact that the object is a singleton is an important part of the API and I'm reluctant to hide that as an implementation detail in the constructor though.  Otherwise you get unexpected behavior when constructing the second instance.

Where, in all the application, would one over construct a second instance? What I was asking for is that the lifetime of this object be scoped to main(). :)

> It would be better to not allow the second instance to be constructed at all, by making the constructor private, and accessing construction through a static method.  That eliminates the possibility of misuse.

OSP_DCHECK() would also eliminate the possibility of misuse.

My position is that we should use the language structure, not static function calls, to manage the lifetime of objects and their connections to the object graph. "Resource Acquisition is Initialization" is a recurrent theme in Chromium code.

There are larger issues with the design of all this, and I don't feel it's productive to discuss them piecemeal in code reviews anymore. I'll follow-up in e-mail...

Patch-set: 1
Reviewer: Gerrit User 1002076 <1002076@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>