DNS-SD: Conversions From mDNS
This CL provides methods to convert from mDNS entities into DNS-SD
entities, to simplify the Querier and Publisher classes. Layers on
top of this one should not look inside MdnsRecord entities, and should
instead treat them as a black box, relying on methods in this layer
to process and look inside of them.

Change-Id: I32916326b6a62d1c572da18865186d2a9768b569