Cast Streaming Receiver demo: libavcodec decoding + libSDL2 for display.
Fills in the receiver demo app to include decoding/rendering of audio
and video, and play it out via a simple GUI.

In doing this, it became clear the Receiver::Consumer API needed to
change to avoid having the client-side code do weird things (like
polling). So, this change includes the necessary Receiver changes to
support that.

The SDL-output code is only activated when the following GN args are set
to true: have_sdl_for_demo_apps, have_ffmpeg_for_demo_apps. Otherwise, a
"dummy player" implementation will be compiled-in that only makes a
short log entry for each frame received, for basic verification of an
end-to-end system.

Bug: openscreen:56
Change-Id: I488547a09699fbc7efaca1a07deaed9b69fc4d28