Cast Streaming Receiver demo: libavcodec decoding + libSDL2 for display.
Fills in the receiver demo app to include decoding/rendering of video,
and display it on-screen. Much of the libavcodec and libSDL2 glue code
here will be reusable when audio is added (coming soon!).

The code is only activated when the following GN args are set to true:
have_sdl_for_demo_apps, have_ffmpeg_for_demo_apps. Otherwise, a "dummy
player" implementation will be compiled-in that only makes a short log
entry for each frame received, for basic verification of an end-to-end
system.

Bug: openscreen:56
Change-Id: I488547a09699fbc7efaca1a07deaed9b69fc4d28