Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4: Code-Review+1
> 
> (1 comment)
> 
> Thank you for putting this change together! It makes the code much cleaner!
> 
> After looking this all through, what are your thoughts on adding a 3rd level of namespaces for mdns vs dnssd?
> IMHO it would make the code more readable, but I'm happy with either approach

I'm not totally sold on it.  If we added an mdns namespace we would see mdns::MdnsRecord in the dnssd code, which is redundant.  If we converted that to mdns::Record, then you see just Record in the mdns code, which is a little too terse for me.

I think this strikes a reasonable balance, and extra namespaces should happen as we encounter genuine name collisions and/or ambiguity.

> Also, let's get this in before I commit any more of my DNS-SD changes :) Otherwise things will keep changing

Patch-set: 4
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>