Update patch set 1
Patch Set 1:

This is going to be nontrivial to land because our fork of chromium_quic is not compatible with the current BoringSSL API:

../third_party/chromium_quic/src/net/third_party/quic/core/crypto/scoped_evp_aead_ctx.cc:11:8: error: no member named 'aead_state' in 'evp_aead_ctx_st'
  ctx_.aead_state = nullptr;

Rolling chromium_quic is nontrivial and the path forward for QUIC to become a quiche embedder.  Unless this is blocking libcast, going to abandon rolling borignssl for now.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0