Update patch set 2
Patch Set 2:

(1 comment)

This makes sense to me.

I believe that one issue of correctness is that the embedder should not instantiate multiple TaskRunners as that could affect ordering guarantees between tasks.  We will now be requiring the embedder to implement the TaskRunner properly, but that seems reasonable if it's well documented.

This might be a good time to start a document for embedders on how to go about implementing the platform API.

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>