Start using ReceiverConfig, ReceiverSession
This patch begins the process of creating a receiver session class that
is integrated with the Receiver implementation. This first step includes
enhancements to the ReceiverConfig class that make it usable by the
session and the receiver, and contains an API-only declaration of
ReceiverSession. Finally, the Receiver class is modified to use the
ReceiverConfig, giving us a base for a fleshed out ReceiverSession
class.

Change-Id: I7585622b95227c340cbdeb05f1ea2d3511bbcf30