[Open Screen] Simplify TxtRecordRdata constructors.
A TXT record is fundamentally a list of variable-length octet data entries.

This updates the main TxtRecordRdata ctor to accept that data type, and removes
others, which were redundant or only used in unit tests.

Adds a convenience function for constructing a TXT record from a list of
strings to mdns_test_util for unit tests.

Bug: openscreen:78
Change-Id: Ifb4b90eb0d359526431eeae7b8091adedd9cc6d3