DNS-SD: Factory Method APIs
This CL defines the DnsSdService class, which is the embedder-specific
way of accessing a Publisher and Querier class. The embedder may make
one of these return nullptr if only one of these is supported, but
having a single class for both makes creating shared resources (ie,
the MdnsService class shared by our Querier and Publisher) simpler.
Additionally, a factory method is defined on this class

Additionally, a factory method is defined for MdnsService, as this is
the primary thing that DnsSdService will need to create

These methods can both be implemented in later patches.

Change-Id: Ica2afcdcbcaba31f0e846e1e179fd78b7fef3679