Move platform/api/trace_logging.h → util/trace_logging.h, etc...
This change moves the parts of trace logging that are library code (and
not platform specific) into util or platform/base, as appropriate.
Library code using the tracing macros is pointed to
util/trace_logging.h. Overall, this "minifies" the platform/api an
embedder must implement to just the TraceLoggerPlatform interface.

Fixed a few C++ style issues found in code as it was moved around.

This change also fixes unit test crashing due to dangling pointers to
detsroyed MockLoggingPlatforms being left around (the global pointer was
not being nulled-out after the tracing unit tests were executed). To
simplify, the ctor/dtor now automatically set/clear the global pointer.
The crashing seems not to have been discovered because the BUILD.gn arg
"enable_trace_logging" defaults to false and so our bots were not
testing things. Will revisit this in a later change...

Bug: openscreen:77
Change-Id: I638a3998af0a3aacf4d8b8e0f059b331582313cb