DNS-SD: Update API for Embedder Use-Cases
This CL applies the API updates discussed to make the API more easily
implementable for embedders

The methodology here is to remove the 'domain' parameter from both
publisher and querier methods because many embedders (such as Android
and Chromium) do not allow the setting of this parameter. Under the
hood, these parameters are still used when examining mDNS records
received off the network, as records with the same instance and
service ids can feasibly be received from multiple DNS domains, so this
parameter would be needed to differentiate them.

Additionally, the Publish API in publisher has been updated to return
an error,so that if the embedder does not support publishing all valid
mDNS records to-spec (as is the case with Android), an error can be
returned here.

Change-Id: I3e3f8562f7ddecfc9fe1999b956e051d93a68904