Move protos to their own BUILD.gn files
Currently, one of the Windows bots is trying to build platform/ and
util/, because it is building all targets that gn knows about, including
ones that are not listed as dependencies, not necessary for any top
level targets, and do not compile on that operating system.

This patch moves protos to their own BUILD.gn files, in an attempt to
keep the Windows bot from discovering unbuildable targets.

Change-Id: I8936812017d8b2fcf82fa922a68a93c645093f2f