Update patch set 2
Patch Set 2:

Isn't this just kicking the can down the road for when we replace more sender pieces?  Or are we always going to be under a kind of is_linux cover?  (In which case this would just follow that pattern.)  Is this solely to avoid the proto duplication to avoid divergence?

Patch-set: 2
Reviewer: Gerrit User 1215804 <1215804@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>