mDNS: Responder + Publisher API
This CL adds 2 new APIs:
- mDNS Responder
- mDNS Publisher

the mDNS Responder API is used to publish new records to our
mDNS implementation. This is based on an API originally
defined by Max in CL:
https://chromium-review.googlesource.com/c/openscreen/+/1847209

mDNS Publisher allows for us to register/deregister/update
records we are currently advertising.

These 2 classes could be combined (as in iteration 2), but both
are relatively complex and have relatively disjoint purposes,
so there is no reason to combine them

Change-Id: Ie39b1405434f5c210875a0e05468d41eca4342b2