Move SerialDeletePtr to util, clean-up ScopedWakeLock platform/api.
Simplifies the ScopedWakeLock platform API to just be a RAII-style
scoped object, managed with a std::unique_ptr<>, and thus leaving the
threading/sequencing details to the embedder. This frees-up
SerialDeletePtr to become a util class, rather than a platform/base
construct.

Added comments to the platform API to explain its purpose and usage.

Bug: openscreen:77, openscreen:89
Change-Id: I242f97a2ca786f51c56adb131b39d82f6e0b5720