mDNS: Publisher Impl - Registration flow
This CL is the first part of the MdnsPublisher's implementation,
specifically it is responsible for handling registration of mDNS Records
through 3 methods:
 - RegisterRecord
 - UpdateRegisteredRecord
 - UnregisterRecord

The implementations of the above are split amongst PTR records and
non-PTR records, because PTR and Non-PTR records are simpler to handle
separately. PTR records are not unique records, while non-PTR records
are.

The API for claiming exclusive ownership is not currently available, so
it is not used here. For that reason, the IsExclusiveOwner method is
mocked. Once that flow has been added, additional tests can be added to
validate that flow.

Bug: openscreen: 57
Change-Id: I5a12fe859e61f09d6bfad8b2072a14ee1cf5326c