mDNS: Responder Bug Fix
Made a quick fix for more generalized mDNS SRV record formats.
In our service, all mDNS SRV records should point to the same instance
where the SRV record is located, which is how the MdnsResponder is
currently coded. In general, this doesn't need to be true. I have
updated the code to support this more general form.

Change-Id: I07cecd124c1b1784037f37377b7271ad287c9acd