Remove Json(Reader|Writer) in favor of namespace
In a recent code review, it was determined that we generally don't want
classes to store JsonWriter and JsonReader instances. Thus, there isn't
really a point in having instances of these classes.

This patch deletes both JsonReader and JsonWriter in favor of a new,
proper json:: namespace in json_serialization.h that includes both
json::Read and json::Write.

Change-Id: Ib31c84339f1bf18919764fd0297a60d34cb66573