Update Receiver ownership
Currently, the embedder is responsible for owning Receivers, but they
are not guaranteed to remain valid and are deleted upon renegotiation.
To avoid confusing ownership and lifetimes, this patch changes
ConfiguredReceivers to use dumb pointers, and have the lifetime of both
the audio and video receivers managed directly by the ReceiverSession.

A new OnReceiversDestroyed event is created, and the unit tests are
updated to ensure it is called appropriately.

Change-Id: I0c4fef00ecbafcc4e9b9abf5af5b507b38fece4b