mDNS: Batch Announce + Goodbye Records
This CL adds a method to batch announce and goodbye records sent from
the MdnsPublisher. It uses a static vector to store the records, and
periodically pulls them all out and sends them together as a single
MdnsMessage. A Static vector is used so that multiple instances of
MdnsPublisher can send records in the same batch.

Bug: openscreen:57
Change-Id: Ibc78a9c239ee5bd7323d72d59c11963f7482d672