mDNS: Probe Skeleton
This is the basic skeleton for a probing API, to handle the mDNS probing
phase. For the near future, the probing implementation will assume that
we are able to successfully acquire any domain name requested without
checking over-the-wire. In future, this can be changed to implement the
full probing protocol.

NOTE: Unit Tests will come once this is checked in. Without
MdnsPublisher further along, there isn't enough code to write UTs.
Change-Id: I0872f9763c14ce64958237f1007d7b396657b109