mDNS: Responder Known Answer Suppression
This CL adds known answer suppression to the mDNS responder. This allows
the querier to send the set of known answers as part of the message, so
that the responder will not send back these answers.

NOTE: This CL does not include support for truncated messages. This will
come in a future CL.
Change-Id: I97af4491a8ca99fcc94b7f137d0120c1054f9752