mDNS: Probe Skeleton Integration
Mark asked to create a follow-up CL to discuss integration between
MdnsProbeManager and the other classes during the review here:
https://chromium-review.googlesource.com/c/openscreen/+/1941078

Here, the following changes are made:
- An API is defined in mdns_probe_manager.h which is implemented by
  MdnsProbeManager.
- MdnsResponder and MdnsPublisher are updated to take in an instance of
  this new interface.
- Now unneeded code has been removed.

A new interface is used so that Unit Tests can use mock
implementations. As MdnsProbeManager is relatively complex, and its
complexity will continue to increase with subsequent CLs to implement
the Probing phase, forcing MdnsResponder and MdnsPublisher unittests
to depend on this class will add unnecessary complexity.

NOTE: Even with MdnsProbeManager::IsProbeQuery() returning false,
      probe queries from other mDNS devices will work as expected.
      These are unicast messages, so MdnsResponder will immediately
      respond with any known answers to those queries, as is expected
      for an mDNS Probe query.
      The reason this needs to be part of the MdnsProbeManager class
      is to allow for conflict resolution when multiple mDNS
      instances are trying to claim the same name, as described in
      the RFC.

NOTE: This CL does not currently contain the necessary UnitTest
      changes. I will update this CL to include those changes once
      the requisite code changes have been examined.

Change-Id: I835f4db07a7ed6d9b9a71059d0278cd0faf7e75e