Add ARM buildbot, update configurations
This patch updates the build bots to use the most appropriate dimensions
for the fleet at this time (Dec 2019), and adds a new ARM64 buildbot.

Change-Id: I5c44619196654aa3f17bbc3519aa00344d5c6de5

Add ARM buildbot, update configurations

This patch updates the build bots to use the most appropriate dimensions
for the fleet at this time (Dec 2019), and adds a new ARM64 buildbot.

Change-Id: I5c44619196654aa3f17bbc3519aa00344d5c6de5