Update patch set 1
Patch Set 1: Code-Review+1

lgtm, but it looks like this failed on the Chromium advisory bots.

If make_crl_tests is not a gtest/unittests target, perhaps it needs to be in one of the "if (!build_with_chromium) {...}" sections, like for the demo apps?

FWIW, we should probably ask Brandon what make_crl_tests is all about, and whether it should be converted to a proper gtest target.

Patch-set: 1
Reviewer: Gerrit User 1002076 <1002076@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1