DNS-SD: Integrate mDNS Service Class Changes
This CL updates the DNS-SD layer's changes to integrate the mDNS layer's
updates to the MdnsService class. The largest change is updating the
publisher class to use the mDNS Probe flow, rather than just publishing
the records it is given.

To do this, a map with key of type DnsSdInstanceRecord is used, which
requires operators defined on a number of classes

Change-Id: Ic955b749616773a52e54259f4848f5740ce65718