Use separate toolchains for cross-compiling
This change makes cross-compiling targets use different toolchains,
instead of only adding different compiler flags.  This has the advantage
that it's easier to have cross-toolchain dependencies like protobuf
object files for the target architecture depending on the protoc
compiler for the host architecture.  This is also how Chromium manages
toolchains.

Bug: openscreen:18
Change-Id: If2547c8cfeec3142db34c2dfae2b93982d8472c5