Add (Video|Audio)Stream to ConfiguredReceivers
Currently, embedders have no way of knowing what stream that libcast
selected, and the SessionConfig class we expose only has a limited set
of fields available, and is missing several important fields, such as
codec.

This patch exposes the AudioStream and VideoStream that were selected,
if any, from the initial offer message on the OnNegotiated callback.

Change-Id: Ia276fb991618a3fb4f8a4cc8c4d77192446397ac