mDNS: MdnsProbeManager Implementation
This CL implements the mDNS Probe Manager class. The implementation here
is based on that described in RFC 6762 section 8, but it diverges
slightly. According to the RFC, probing should be done with all records
which are going to be published by the service. Here, only one of the A
or AAAA records to be published is used. Beyond this, the probe phase
respects the RFC.

The process used, in summary, is to spin up a new Probe when the user
attempts to claim a given domain. If this probe succeeds, the domain is
marked as claimed by this service, and the provided callback is
informed that the domain has been claimed. any incoming probe queries
for this domain will receive a response of an A or AAAA record
associated with this domain. If the probe fails, the domain is
automatically modified (as suggested in RFC 6763 appendix D) and
probing begins again with this modified domain.

In the case where a probe request is received for the same domain which
is already being probed for, a decision is reached on whether to accept
the other host's claim to this name or the current machine's based on
comparison of the to-be-published records provided in each query.

For further details on the probing phase, see
https://tools.ietf.org/html/rfc6762#section-8

Change-Id: Iac35a8819db0caab6b323918a9012cf6daee4604