Temporarily unblock bots
This patch gets the checkin system working again by adding a
clang-format binary for the bots. Currently, the bot is too dumb to run
gclient sync and pick up the clang binary, and we don't own the script
so I'm not sure what the longterm fix will be.

TBR: miu@chromium.org
Change-Id: Ic75d7ed89d2c7fb4e6056b3248a2f2a3ca7050bd