Discovery E2E Test
This CL creates an E2E test of the discovery code in discovery/ and
cast/common/discovery. It makes use of the Discovery changes previously
introduced as part of CL:
https://chromium-review.googlesource.com/c/openscreen/+/2008504

In the test, service publication and discovery will both be started.
When discovery sees a new service instance, it will log that to the
console.

Change-Id: I05eddcf18a335180d8918db048593ecab7b46c50