Update patch set 4
Patch Set 4:

@mfoltz

I've been thinking more about what you suggested for creating an initialization parameters struct rather than passing in the NetworkInterface - and I think this is a better approach
In addition to the network interface, we have some currently hard-coded values that should be changed to configurable (for example, how many times to announce newly published records)

If possible, I'd prefer to do this as a follow-up CL once this one is in (ideally with the current NetworkInterface approach, to simplify the follow-up patch). It would touch a lot of files not included in the current CL, so a separate CL should improve review-ability. WDYT?

Patch-set: 4
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0