Update patch set 1
Patch Set 1: Commit-Queue+1

Is there any way to run this against a swarming bot without checking in and rolling? 

It appears that the data tag is used elsewhere for the same type of dependencies: 

https://osscs.corp.google.com/chromium/chromium/src/+/master:cc/BUILD.gn;l=747?q=data%20file:BUILD.gn

But I'm not sure if that's enough for swarming, or if we will have to have Chromium-only data bundle tags (bleck).

Patch-set: 1
Reviewer: Gerrit User 1323871 <1323871@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false