Update patch set 4
Patch Set 4:

(4 comments)

Replied to a few comments that I would like to clarify further
Per below comment, going to put further work on this CL on hold pending further discussion

> This is a good patch for further discussion.  We can go one of two routes:

I've added some of my thoughts on the two approaches you mentioned as well as an option 3, but let me know if you'd rather I set up a meeting for Monday

I have a slight preference for option 1, just because it's simpler for our embedders to use and harder to get wrong. IMHO the embedder shouldn't have to understand our task runner - they should treat it as a black-box to pass to our methods. But I'm happy to go either way

> 1. Allow clients to call some of our APIs on a thread of their choosing, and we handle thread safety by moving work to the appropriate thread.
> 
> Pro: Simpler from the client's POV.
> Cons: Simple getters/setters are challenging, requires mutexes or careful coding.
>       Can't support very many synchronous APIs.
>       Lots of boilerplate "trampoline" code behind our public APIs like what's included here.

A bit of further context:
In our current discovery API, we require very few getters/setters. The only case where it's needed is the GetServices() call in discovery/public
The two publicly visible DNS-SD classes (DnsSdPublisher and DnsSdQuerier) do not provide any getters/setters, and instead rely on callbacks

So we would only require one lock for all of discovery, and it would only get locked pretty rarely (how often do we expect the caller to call GetServices rather than relying on the callbacks?).
Though it's quite possible our users will abuse this function and cause deadlock or poor performance
("We want the list to always be up to date, so we should call the GetServices method 100x per second and update our UX every time, right?")

Regarding synchronous APIs, discovery doesn't currently support any of those at the mDNS:
- The Querier starts up an synchronous method, with callbacks as services are found
- Publication starts up a new mDNS probe query, publishing new services and letting the embedder know once this succeeds

Given this, our only 'synchronous' parts of the API are those that occur at the discovery/public and discovery/dnssd layers, before we hit mDNS. The only 2 cases for this are:
- Registration errors (in the DNS-SD layer's publication flow)
- Type conversion errors (in the discovery/public layer, between ServiceInfo and DnsSdInstanceRecord)

These seems like good candidates for custom error types and the reporting API I checked in earlier today (with a new OnRecoverableError() callback, or similar)

> 2. Require clients to call everything on the thread they're using from the task runner.
> 
> Pro: Simpler, and harder to get wrong.
> Con: Clients will have to write some code to bounce work to the OSL thread when necessary (i.e. from the UI or IPC handlers).

This boilerplate bounding code is actually really minimal. I wrote the E2E test originally using this approach, and it only involved the following:

  task_runner_ptr->PostTask([&receiver]() { receiver.StartDiscovery(); });
  task_runner_ptr->PostTask([&publisher, &v4]() { publisher.Register(v4); });
  task_runner_ptr->PostTask([&publisher, &v6]() { publisher.Register(v6); });
  task_runner_ptr->PostTask(
      [&publisher, &v4_and_v6]() { publisher.Register(v4_and_v6); });

It gets a bit more complex to handle errors on the Register methods, but not by much. Though it does introduce threading concerns.

Additionally, thread ownership of objects shouldn't be a worry, due to the change here to make the DnsSdService factory return a SerialDeletePtr - all dtors are now guaranteed to run on the Task Runner thread



There is also an option 3 - kind of a hybrid of the above 2.
- We bounce all calls to the TaskRunner(option 1)
- We require all getters/setters be run on the Task Runner thread (option 2)

Then, we can get away without any locks, and still have nearly all caller code require no task-runner interaction


WDYT?

Patch-set: 4