Improve raspberry pi documentation
Currently, the docs state that you must have a compatible version of
libavcodec and libsdl2, however developers have no way of knowing what
that means. This patch adds a example command that can be run on a
raspberry pi to pull down the exact versions of these packages used to
generate the sysroot image, ensuring that people can actually use the
sysroot images provided.

Change-Id: I28626ce92ca6903b00bff3146877cd0576b1fe24