mDNS: Add Network Interface to A and AAAA records
This CL adds the Network Interface index to A and AAAA records. This is
not intended to support any new functionality, but only to add the
minimum functionality needed to split the mDNS and DNS-SD layer
modifications needed to fix the attached bug.

NOTE: The following will be changed in a follow-up test:
- Unit Tests for new A and AAAA record comparison functionality.
- Default argument for NetworkInterfacIndex in A and AAAA records will
  be removed.

Bug: openscreen:105
Change-Id: I79d8b0e45468e1b219bff770f68e859c3035d281