Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > (7 comments)
> > 
> > Can you explain a bit more how this works as an E2E test?
> > 
> > It looks like the binary publishes a hard coded set of services.  If you were to start it on two distinct machines, the service names would conflict and probing/conflict resolution would kick in, so you'd get nondeterministic results based on IP addresses (I think, based on lexicographic ordering of A/AAAA).
> > 
> > I expected something more like the following:
> > 
> > - The binary could be started as a publisher, listener, or both
> > - The binary gets passed a list of services to publish from command-line flags or config file
> > - The binary gets passed a list of services to expect to receive by the same mechanism
> > - Binary exits successfully when all expected services are published & received
> > - Some of the parameters like the retry/timeout could also be configured
> > 
> > Maybe some of the test configuration/validation is coming in later patches?
> 
> Will reply to comments after
> 
> But at a high level, this is intended to be useful purely as a E2E test from a single machine (over the loopback interface), so it could be used as a trybot or similar. Per our previous discussions, a separate Demo will be written after this is in, based on similar code (once further validations are added to this E2E test) to test between two machines

Ok.  Some comments at the beginning of main.cc would help explain how it was to be used.  If this is for a test, it's confusing to have it in a folder named "demo."

Also, you'll want to pass in some config to have it use the loopback interface exclusively; otherwise other mDNS endpoints on the same LAN could affect the outcome of the test.  (Even at that, if two copies of the test run on the same machine and both use loopback, you could get different results.)

Patch-set: 5
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>