mDNS: Change Continuous Multicast Query Handling
When querying in mDNS, a number of mDNS records will be returned, each
of which will have a TTL. According to RFC 6762 section 5.2, when an
mDNS Record is nearing expiry, that record should be refreshed.
Currently, this implementation does that by sending an mDNS query for
each record individually and using that to update the record. Per the
RFC, it should instead be sending the original query/queries to refresh
near-expiry records. This CL accomplishes that change.

This alternative querying format should greatly reduce the number of
queries needed (one query for all PTR records instead of one per
record, and all records from a host (A, AAAA, TXT, SRV) can be
refreshed with a single query rather than individually. Additionally,
this change paves the way for NSEC record negative response records to
further reduce traffic.

Change-Id: I6601e936a43ccd0113fc40b95b9204d7c200ddfb