Clarify Alarm::Schedule() API with kImmediately.
Alarm::Schedule() was often being call with a {} time_point argument, to
request the task be run immediately. Instead, add a kImmediately
constant to make it clear to readers of the calling code what is
intended.

This change also makes minor changes inside the Alarm::Schedule()
implementation to prevent underflow in the delay calculations when
kImmediately is used.

Change-Id: I22e6f0d5ac4c1fd6926c8d20dd5332acd538ca68