Discovery: Fix API boundary threading models
This CL updates the callbacks used to push data from lower layers
(mDNS -> DNS-SD, DNS-SD -> public, and public -> embedder) to be called
on the task runner instead of in the same task as the higher layers.
Previously, it was possible for higher layer callbacks to invalidate
iterators and references used by the lower layer, leading to hard-to-
find bugs.
A follow-up CL will address changing these callback calls to use
WeakPtrs instead of pointers directly

In the process, callbacks were also updated to pass by value instead of
by reference

Additionally, a small E2E test modification was added which exposed
this issue.

Change-Id: Iaa5edfcf18fb21548a728a921117e141656c81ba