Update patch set 6
Patch Set 6:

(6 comments)

All of this code should already be running on the same sequence as the task runner, are these changes to enforce sequencing of operations?

Can you provide an example of the issue you found with iteration?  Sometimes the solution is to make a copy of a container or some other flag to prevent re-entrancy when reading/mutating the same container.

Patch-set: 6
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>