Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (8 comments)
> 
> > Patch Set 5:
> > 
> > (9 comments)
> > 
> > Could you please give an example of such an invalidation performed by an upper layer?
> 
> The place I originally found this was in mdns_querier:500
> The OnRecordChanged callback would fire kExpired for a PTR record. This would make dnssd/impl/querier_impl stop the mDNS kALL query associated with the pointed to domain. This would then call mDNS Querier and update |callbacks_| which would invalidate the iterator still being used by the original call, resulting in an infinite loop since the exit condition is never hit

Right, the original call could make a copy of the callback pointers if a callback could modify the list.  Or it could set a flag so the callbacks list is not modified inside a callback.

IMO, that would be easier to reason about than using the task runner to create implicit sequencing of mutation operations.

Patch-set: 6
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>