Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > > Patch Set 6:
> > > 
> > > (8 comments)
> > > 
> > > > Patch Set 5:
> > > > 
> > > > (9 comments)
> > > > 
> > > > Could you please give an example of such an invalidation performed by an upper layer?
> > > 
> > > The place I originally found this was in mdns_querier:500
> > > The OnRecordChanged callback would fire kExpired for a PTR record. This would make dnssd/impl/querier_impl stop the mDNS kALL query associated with the pointed to domain. This would then call mDNS Querier and update |callbacks_| which would invalidate the iterator still being used by the original call, resulting in an infinite loop since the exit condition is never hit
> > 
> > Right, the original call could make a copy of the callback pointers if a callback could modify the list.  Or it could set a flag so the callbacks list is not modified inside a callback.
> > 
> > IMO, that would be easier to reason about than using the task runner to create implicit sequencing of mutation operations.
> 
> If that's your preference, LMK and I will put up a patch. IMHO the first option here seems to be the cleaner of the two, so I'd prefer to go that route
> 
> My concern that lead to this approach is that these bugs are incredibly hard to find - This above example I only found because the E2E tests happened to stumble across them, and if I didn't have a consistent repro it would have been near impossible to find. If this happened once we release the library, it would be even harder.
> With this approach, we create a contract that on a given task runner call, the higher layers can call into the lower layers, but lower never calls into higher - so these bugs are no longer possible.
> 
> At the mDNS Layer, it's possible there are other edge cases that haven't been considered. Especially given this layer's current complexity which will only increase with time. If we eventually introduce further callback flows (we've discussed doing this before), it's possible similar bugs would be introduced. My goal is to avoid that
> At the DNS-SD + public layers, we don't know exactly how our library will be used, and I fully expect that some embedders will do things we never expect. These changes prevent against similar bugs arising from those use cases
> 
> WDYT?

TL;DR I think this is an API contract issue, not really a "when to run the callback" issue.

The core issue here, in my opinion, are the contracts between the layers in the discovery module, especially what callbacks are allowed to do.

If DNS-SD can pass a callback into mDNS that mutates mDNS state, then you'll run the risk of that callback doing things that you don't expect.  If you move that callback to a task runner, you'll still run that risk, but the task runner will make the exact sequence of operations less deterministic (since there may be other pending tasks that could turn change the final state).

The safest thing to do, is to restrict the behavior of mDNS callbacks to only take POD/structs, or const pointers/refs to stateful objects in the system, so the state mDNS can'be mutated in the callback.

If there are specific mutations that are necessary, then maybe have the callback return a value describing the mutation so that mDNS can apply it after running the callback.

Patch-set: 6
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>